+++ be/src/test/java/codesquad/bookkbookk/chapter/integration/ChapterTest.java
@@ -24,6 +24,7 @@ import codesquad.bookkbookk.domain.chapter.data.entity.Chapter;
 import codesquad.bookkbookk.domain.chapter.repository.ChapterRepository;
 import codesquad.bookkbookk.domain.member.data.entity.Member;
 import codesquad.bookkbookk.domain.member.repository.MemberRepository;
+import codesquad.bookkbookk.domain.topic.data.entity.Topic;
 import codesquad.bookkbookk.domain.topic.repository.TopicRepository;
 import codesquad.bookkbookk.util.TestDataFactory;
 
@@ -155,6 +156,44 @@ public class ChapterTest extends IntegrationTest {
             softAssertions.assertThatThrownBy(() -> chapterRepository.findById(chapter.getId())
                             .orElseThrow(ChapterNotFoundException::new))
                     .isInstanceOf(ChapterNotFoundException.class);
+        });
+    }
+
+    @DisplayName("성공적으로 토픽을 조회한다")
+    @Test
+    void readTopicList(){
+        //given
+        Member member = TestDataFactory.createMember();
+        memberRepository.save(member);
+        String accessToken = jwtProvider.createAccessToken(member.getId());
+
+        BookClub bookClub = TestDataFactory.createBookClub();
+        bookClubRepository.save(bookClub);
+
+        Book book = TestDataFactory.createBook1(bookClub);
+        bookRepository.save(book);
+
+        Chapter chapter = new Chapter(book, "first");
+        chapterRepository.save(chapter);
+
+        Topic topic1 = new Topic(chapter, "토픽1");
+        Topic topic2 = new Topic(chapter, "토픽2");
+        topicRepository.save(topic1);
+        topicRepository.save(topic2);
+
+        //when
+        ExtractableResponse<Response> response = RestAssured
+                .given().log().all()
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
+                .when()
+                .get("/api/chapters/" + chapter.getId() + "/topics")
+                .then().log().all()
+                .extract();
+
+        //then
+        SoftAssertions.assertSoftly(softAssertions -> {
+            softAssertions.assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+            softAssertions.assertThat(response.jsonPath().getList("")).hasSize(2);
         });
     }
